+++ x64_dbg_dbg/memory.h
@@ -4,7 +4,7 @@
 #include "_global.h"
 
 uint memfindbaseaddr(HANDLE hProcess, uint addr, uint* size);
-bool memread(HANDLE hProcess, const void* lpBaseAddress, void* lpBuffer, DWORD nSize, SIZE_T* lpNumberOfBytesRead);
+bool memread(HANDLE hProcess, const void* lpBaseAddress, void* lpBuffer, SIZE_T nSize, SIZE_T* lpNumberOfBytesRead);
 void* memalloc(HANDLE hProcess, uint addr, DWORD size, DWORD fdProtect);
 
 #endif // _MEMORY_H
